+++ CHANGELOG.md
@@ -52,6 +52,8 @@ and this project adheres to
   - [#3956](https://github.com/bpftrace/bpftrace/pull/3956)
 - Add signed type checking for map assignments
   - [#4132](https://github.com/bpftrace/bpftrace/pull/4132)
+- Add signed type checking for map keys
+  - [#4136](https://github.com/bpftrace/bpftrace/pull/4136)
 #### Deprecated
 #### Removed
 #### Fixed